refactored test classes and packages to have the same structure as codebase
