--- src/test/java/org/finra/msd/sparkcompare/baseclasses/BaseJunitForSparkCompare.java
@@ -1,29 +1,0 @@
-package org.finra.msd.sparkcompare.baseclasses;
-
-import org.finra.msd.memorydb.MemoryDbHsql;
-import org.finra.msd.sparkfactory.SparkFactory;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-
-public class BaseJunitForSparkCompare {
-
-    protected final String outputDirectory = System.getProperty("user.dir") + "/sparkOutputDirectory";
-
-    @BeforeClass
-    public static void setUpClass() {
-        SparkFactory.initializeSparkLocalMode("local[*]" , "WARN" ,"1");
-
-        if (MemoryDbHsql.getInstance().getState() != 1 )
-        {
-            MemoryDbHsql.getInstance().initializeMemoryDB();
-            MemoryDbHsql.getInstance().stageTablesAndTestData();
-        }
-    }
-
-    @AfterClass
-    public static void tearDownClass() {
-        SparkFactory.stopSparkContext();
-        //TODO: need to implement a better way to close in memory DB
-        //MemoryDbHsql.getInstance().shutdownMemoryDb();
-    }
-}
